requirements.txt
streamlit
pandas
python-pptx
openpyxl
fpdf
